press float

+++ boatBoot/src/main/java/com/wisely/dao/BigDemoMetadataDao.java
@@ -11,6 +11,6 @@ public interface BigDemoMetadataDao extends CustomRepository<BigDemoMetadata, St
 
 	
 	@Query("select count(meta), meta.pk from BigDemoMetadata meta where meta.sampleName=?1 and meta.temparture=?2 and meta.press=?3 and meta.testModelName=?4 and meta.testSystemName=?5")
-	List<Object> CountBySamplenameAndTempartureAndPressAndTestNames(String SmapleName,float Temparture ,int press,String testMname,String testSname);
+	List<Object> CountBySamplenameAndTempartureAndPressAndTestNames(String SmapleName,float Temparture ,float press,String testMname,String testSname);
 
 }
